Add ability to swipe weeks at teacher lessons screen
