+++ app/src/main/java/ru/neexol/rtut/presentation/screens/teacher/ui/TeacherScreen.kt
@@ -7,10 +7,7 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material.ExperimentalMaterialApi
 import androidx.compose.material.SnackbarHost
 import androidx.compose.material.SnackbarHostState
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
@@ -19,6 +16,8 @@ import com.google.accompanist.pager.ExperimentalPagerApi
 import com.google.accompanist.pager.rememberPagerState
 import kotlinx.coroutines.launch
 import ru.neexol.rtut.R
+import ru.neexol.rtut.presentation.components.scrollingPair
+import ru.neexol.rtut.presentation.components.syncScroll
 import ru.neexol.rtut.presentation.screens.teacher.TeacherViewModel
 
 @ExperimentalAnimationApi
@@ -29,6 +28,12 @@ fun TeacherScreen(vm: TeacherViewModel = hiltViewModel()) {
 	val coroutineScope = rememberCoroutineScope()
 	val snackbarHostState = remember { SnackbarHostState() }
 	val weekPagerState = rememberPagerState(vm.dayWeek.second.coerceAtMost(15))
+	val lessonsPagerState = rememberPagerState(vm.dayWeek.second.coerceAtMost(15))
+
+	val scrollingPair by remember {
+		derivedStateOf { scrollingPair(weekPagerState, lessonsPagerState) }
+	}
+	LaunchedEffect(scrollingPair) { syncScroll(scrollingPair) }
 
 	LaunchedEffect(vm.uiState.message) {
 		vm.uiState.message?.let {
@@ -50,9 +55,9 @@ fun TeacherScreen(vm: TeacherViewModel = hiltViewModel()) {
 		}
 		FindTeacherBar(vm)
 		Box(Modifier.fillMaxSize()) {
-			LessonsList(
+			LessonsPager(
+				state = lessonsPagerState,
 				uiState = vm.uiState,
-				week = weekPagerState.currentPage,
 				onClassroomCopy = {
 					coroutineScope.launch {
 						snackbarHostState.currentSnackbarData?.dismiss()
